Swagger documentation + minor fix v1.0

+++ routes/user.route.js
@@ -27,7 +27,7 @@ router.post('/Login', validate(loginSchema), async (req, res) => {
   const data = req.body
   const ret = await db('Users').where('username', data.username)
   if (ret.length === 0) {
-    return res.status('200').json({ message: 'User not found' })
+    return res.status(401).json({ message: 'Invalid credential' })
   } else {
     const result = await bcrypt.compare(data.password, ret[0].password)
     if (result) {
@@ -79,6 +79,9 @@ router.post('/ResetPassword', async (req, res) => {
     }
   })
   const user = await userModel.findOne({ username: data.username }, userViewModel)
+  if (!user) {
+    return res.status(401).json({ error_message: 'Invalid username' })
+  }
   const otp = otpGenerator.generate(6, { upperCaseAlphabets: false, specialChars: false })
   try {
     const data = fs.readFileSync('./html_template/OTP.html')
@@ -88,7 +91,7 @@ router.post('/ResetPassword', async (req, res) => {
     const mainOptions = {
       from: 'SWEN Bank',
       to: user.email,
-      subject: 'OTP xác nhận giao dịch',
+      subject: 'OTP xác nhận đổi mật khẩu',
       html: htmlString
     }
     transporter.sendMail(mainOptions, async (err, info) => {
@@ -105,6 +108,7 @@ router.post('/ResetPassword', async (req, res) => {
           user_id: user.id
         }
         await resetPasswordOTPModel.add(otpData)
+        res.status(200).json({ message: 'OTP code has been sent!' })
       }
     })
   } catch (err) {
@@ -115,16 +119,19 @@ router.post('/ResetPassword', async (req, res) => {
 router.post('/VerifyOTP', async (req, res) => {
   const data = req.body
   const resetPasswordOTP = await resetPasswordOTPModel.findOne({ otp: data.otp }, resetPasswordOTPViewModel)
-  const user = await userModel.findOne({ id: resetPasswordOTP.id })
+  if (!resetPasswordOTP) {
+    return res.status(401).json({ error_message: 'Invalid OTP' })
+  }
+  const user = await userModel.findOne({ id: resetPasswordOTP.user_id })
   if (new Date(resetPasswordOTP.expired_at).getTime() < new Date().getTime()) {
     await transactionOTPModel.delete(resetPasswordOTP.id)
-    return res.status('200').json({ error_message: 'OTP hết hạn' })
+    return res.status(401).json({ error_message: 'OTP expired!' })
   }
   const salt = await bcrypt.genSalt(10)
   data.password = await bcrypt.hash(data.password, salt)
   user.password = data.password
   await userModel.update(user.id, user, userViewModel)
-  return res.status('200').json({ message: 'Đã thay đổi' })
+  return res.status(200).json({ message: 'Password changed!' })
 })
 
 router.use(currentUserMdw)
@@ -136,7 +143,6 @@ router.patch('/', async (req, res) => {
   // const salt = await bcrypt.genSalt(10)
   // oldUser.password = await bcrypt.hash(data.password, salt)
 
-  delete data.token
   const ret = await userModel.update(currentUser.id, data, userViewModel)
   res.status(201).json(ret[0])
 })
@@ -144,13 +150,13 @@ router.patch('/', async (req, res) => {
 router.post('/', validate(userSchema), async (req, res) => {
   const currentUser = res.locals.currentUser
   if (currentUser.role_id !== 1 && currentUser.role_id !== 3) {
-    return res.status('403').json({ message: 'Không đủ quyền truy cập' })
+    return res.status(403).json({ message: 'Warning: You do not have permission to access the API!' })
   }
   let data = req.body
   const oldUser = await userModel.findOne({ username: data.username }, userViewModel)
 
   if (oldUser) {
-    return res.json('409').json({ message: 'User already exist' })
+    return res.status(409).json({ message: 'User already exist' })
   }
 
   data.role_id = data.role_id || 2
@@ -181,7 +187,7 @@ router.post('/', validate(userSchema), async (req, res) => {
 router.get('/Employees', async (req, res) => {
   const currentUser = res.locals.currentUser
   if (currentUser.role_id !== 1) {
-    return res.status('403').json({ message: 'Không đủ quyền truy cập' })
+    return res.status(403).json({ message: 'Warning: You do not have permission to access the API!' })
   }
 
   const ret = await userModel.fetch({ role_id: 3 }, userViewModel)
@@ -191,7 +197,7 @@ router.get('/Employees', async (req, res) => {
 router.delete('/Employees/:id', async (req, res) => {
   const currentUser = res.locals.currentUser
   if (currentUser.role_id !== 1) {
-    return res.status('403').json({ message: 'Không đủ quyền truy cập' })
+    return res.status(403).json({ message: 'Warning: You do not have permission to access the API!' })
   }
   const id = req.params.id
   const ret = await userModel.delete(id)
